backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
stripe
geopy
pygeohash
requests
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
Pillow>=10.0.0
cloudinary>=1.36.0
